feat: refactor package structure for ride history response classes
