+++ drumigo/src/main/java/com/ftn/drumigo/dto/history/response/DriverRideHistoryItemResponse.java
@@ -1,4 +1,4 @@
-package com.ftn.drumigo.dto;
+package com.ftn.drumigo.dto.history.response;
 
 import java.math.BigDecimal;
 import java.time.Instant;
